requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.34.138
botocore==1.34.138
Django==4.2.13
django-storages==1.14.3
django-taggit==5.0.1
django-unused-media==0.2.2
jmespath==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
